Day 6 part 2

+++ Days/Day6/Model/Race.cs
@@ -2,7 +2,7 @@ namespace AdventOfCode2023.Days.Day6.Model
 {
     public class Race
     {
-        public int Time { get; set; }
-        public int Distance { get; set; }
+        public long Time { get; set; }
+        public long Distance { get; set; }
     }
 }
